requirements.txt
# Core
requests>=2.31.0

# Scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Database
prisma>=0.15.0
psycopg2-binary>=2.9.0

# API
fastapi>=0.111.0
uvicorn>=0.29.0
httpx>=0.27.0

# Dev / lint / test
ruff>=0.5.0
pytest>=8.0.0
pytest-env>=1.0.0
python-dotenv>=1.0.0
